# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Smatch Badminton Backend API
  description: |
    API for managing badminton court information, bookings, payments, and serving vector map tiles.
    
    ## Features
    - **Courts Management**: CRUD operations for badminton courts
    - **Search & Autocomplete**: Real-time search with Vietnamese support
    - **Availability**: Time slot availability with pricing
    - **Bookings**: Court reservation system
    - **Payments**: ZaloPay integration with real-time WebSocket notifications
    - **Map Tiles**: Vector tiles for interactive maps
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
...
security:
  - BearerAuth: []
paths:
...
  /api/courts/{courtId}/availability:
    get:
      tags:
        - Availability
      summary: Get court availability
      description: Get availability for all sub-courts of a court on a specific date
      operationId: getCourtAvailability
      parameters:
        - name: courtId
          in: path
          required: true
          description: Court UUID
          schema:
            type: string
            format: uuid
        - name: date
          in: query
          required: true
          description: Date in YYYY-MM-DD format
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Court availability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityResponse'
        '404':
          description: Court not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/bookings:
    get:
      tags:
        - Bookings
      summary: Get bookings by phone
      description: Get all bookings for a phone number
      operationId: getBookingsByPhone
      parameters:
        - name: phone
          in: query
          required: true
          description: Phone number
          schema:
            type: string
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingListResponse'

    post:
      tags:
        - Bookings
      summary: Create a booking
      description: |
        Create a new court booking. Minimum duration is 1 hour, in 30-minute increments.
        Requires an authenticated Firebase user (registered or anonymous). The `userId`
        is taken from the Bearer token and not provided in the request body.
      operationId: createBooking
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Sub-court not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Time slot already booked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/bookings/{id}:
    get:
      tags:
        - Bookings
      summary: Get booking by ID
      description: Get a specific booking by its UUID
      operationId: getBookingById
      parameters:
        - name: id
          in: path
          required: true
          description: Booking UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Booking details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Bookings
      summary: Cancel a booking
      description: Cancel an existing booking
      operationId: cancelBooking
      parameters:
        - name: id
          in: path
          required: true
          description: Booking UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Booking cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          description: Booking already cancelled or completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              example: Error description
...
    TimeSlot:
      type: object
      properties:
        startTime:
          type: string
          example: '06:00'
        endTime:
          type: string
          example: '06:30'
        isAvailable:
          type: boolean
        price:
          type: integer
          description: Price in VND for the 30-minute slot
    SubCourtAvailability:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        slots:
          type: array
          items:
            $ref: '#/components/schemas/TimeSlot'
    AvailabilityData:
      type: object
      properties:
        courtId:
          type: string
          format: uuid
        courtName:
          type: string
        date:
          type: string
          format: date
        dayType:
          type: string
          enum: [weekday, weekend, holiday]
        openingTime:
          type: string
        closingTime:
          type: string
        subCourts:
          type: array
          items:
            $ref: '#/components/schemas/SubCourtAvailability'
    AvailabilityResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AvailabilityData'

    # Booking Schemas
    CreateBookingRequest:
      type: object
      required:
        - subCourtId
        - guestName
        - guestPhone
        - date
        - startTime
        - endTime
      properties:
        subCourtId:
          type: string
          format: uuid
        guestName:
          type: string
        guestPhone:
          type: string
        guestEmail:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
          example: '10:00'
        endTime:
          type: string
          example: '12:00'
        notes:
          type: string
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        subCourtId:
          type: string
          format: uuid
        subCourtName:
          type: string
        courtId:
          type: string
          format: uuid
        courtName:
          type: string
        guestName:
          type: string
        guestPhone:
          type: string
        guestEmail:
          type: string
          nullable: true
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        totalPrice:
          type: integer
          description: Total price in VND
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    BookingResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Booking'
    BookingListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Booking'

    # Payment Schemas
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: Firebase ID token obtained from Firebase Auth SDK